+++ source/Network/Implementation/PacketFragment.cpp
@@ -68,7 +68,7 @@ PacketFragment::PacketFragment(
             data.reserve(OffsetFragmentPayload + size);
 
             auto inserter = back_inserter(data);
-            Push(inserter, sequence.Value());
+            Push(inserter, static_cast<std::uint16_t>(sequence.Value()));
             data.push_back(fragmentId);
 
             data[OffsetSequence] |= MaskTopByteIsFragmented;
